renamed 'frontend/src/components/forecast' to 'frontend/src/components/city_forecast
