+++ frontend/src/components/city_forecast/forecast_day_card.tsx
